Fix issue #3335 with renamed CSV files
Change External Data mechanism for importing CSV files for search()/pulldata() functionalty so that it no longer has to rename a file to track file changes. The CSV file timestamp is now captured in a metadata table stored in the SQLite database holding the imported file. Updates to the file will cause the database to be rebuilt the next time the form is loaded. If the timestamp is unchanged, we skip importing the data.

Now that the file is no longer being renamed, other ODK functionality that relies on CSV files can operate as expected without having their files disappear from under them. However, the each CSV file will still be imported into a SQLite database by the External Data functionality, since it is not implemented narrowly enough to understand which CSV files are actually referenced by the form via search()/pulldata().

+++ collect_app/src/main/java/org/odk/collect/android/external/ExternalDataReaderImpl.java
@@ -60,11 +60,15 @@ public class ExternalDataReaderImpl implements ExternalDataReader {
         File dbFile = new File(dataSetFile.getParentFile().getAbsolutePath(),
                 dataSetName + ".db");
         if (dbFile.exists()) {
-            // this means the someone updated the csv file, so we need to reload it
-            boolean deleted = dbFile.delete();
-            if (!deleted) {
-                Timber.e("%s has changed but we could not delete the previous DB at %s",
-                        dataSetFile.getName(), dbFile.getAbsolutePath());
+            // Determine if we need to reimport
+            if (ExternalSQLiteOpenHelper.shouldUpdateDBforDataSet(dbFile, dataSetFile)) {
+                boolean deleted = dbFile.delete();
+                if (!deleted) {
+                    Timber.e("%s has changed but we could not delete the previous DB at %s",
+                            dataSetFile.getName(), dbFile.getAbsolutePath());
+                    return true;
+                }
+            } else {
                 return true;
             }
         }
@@ -96,18 +100,6 @@ public class ExternalDataReaderImpl implements ExternalDataReader {
             // then just exit and do not process any other CSVs.
             return false;
 
-        } else {
-            // rename the dataSetFile into "dataSetFile.csv.imported" in order not to be
-            // loaded again
-            File importedFile = new File(dataSetFile.getParentFile(),
-                    dataSetFile.getName() + ".imported");
-            boolean renamed = dataSetFile.renameTo(importedFile);
-            if (!renamed) {
-                Timber.e("%s could not be renamed to be archived. It will be re-imported "
-                        + "again! :(", dataSetFile.getName());
-            } else {
-                Timber.e("%s was renamed to %s", dataSetFile.getName(), importedFile.getName());
-            }
         }
         return true;
     }
